backend/pyproject.toml
[project]
name = "sage-retirement-backend"
version = "0.1.0"
description = "FastAPI backend for AI-powered retirement planning with Azure AI Projects"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "azure-identity>=1.15.0",
    "azure-ai-agents>=1.1.0",
    "azure-core>=1.29.5",
    "openai>=1.30.0",
]

[project.optional-dependencies]
eval = [
    "azure-ai-evaluation>=1.0.0",
]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests", "."]
python_files = ["test_*.py"]
